+++ em-app-as/src/main/java/ca/bigmwaj/emapp/as/dao/platform/AccountDao.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.ArrayList;
 
 @Repository
-public interface AccountDao extends AbstractDao<AccountEntity, Long> {
+public interface AccountDao extends AbstractDao<AccountEntity, Short> {
 
     default Class<AccountEntity> getEntityClass() {
         return AccountEntity.class;
