+++ em-app-as/src/main/java/ca/bigmwaj/emapp/as/dao/platform/UserDao.java
@@ -9,7 +9,7 @@ import java.util.ArrayList;
 import java.util.Optional;
 
 @Repository
-public interface UserDao extends AbstractDao<UserEntity, Long> {
+public interface UserDao extends AbstractDao<UserEntity, Short> {
 
     default Class<UserEntity> getEntityClass() {
         return UserEntity.class;
